requirements.txt
anthropic
beautifulsoup4
html2text
pyyaml
python-dotenv
jinja2
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
